usb: split gamepad into grid_usb_gamepad.c/h
Gamepad is fully independent of keyboard/mouse/macro — no shared state,
separate tx queue, separate connect/disconnect lifecycle. Move it to its
own translation unit. grid_usb_hid now covers only keyboard, mouse, and
the macro sequencer that ties them together.
