+++ common/src/c/grid_usb_hid.c
@@ -14,7 +14,6 @@
 struct grid_macro_model grid_macro_state = {0};
 struct grid_keyboard_model grid_keyboard_state = {0};
 struct grid_mouse_model grid_mouse_state = {0};
-struct grid_gamepad_model grid_gamepad_state = {0};
 
 //--------------------------------------------------------------------+
 // Static helpers — defined before callers to avoid forward declarations
@@ -61,12 +60,6 @@ static int32_t grid_usb_mouse_move(struct grid_mouse_model* usb_mouse, int8_t po
   return 0 == tud_hid_mouse_report(HID_ITF_PROTOCOL_MOUSE, usb_mouse->buttons, delta[0], delta[1], delta[2], 0);
 }
 
-// Axes RX/RY/RZ (3,4,5) are passed in reverse order to match the HID descriptor layout.
-static int32_t grid_usb_gamepad_report_send(struct grid_gamepad_model* usb_gamepad) {
-  return 0 == tud_hid_gamepad_report(3, usb_gamepad->axis[0], usb_gamepad->axis[1], usb_gamepad->axis[2], usb_gamepad->axis[5], usb_gamepad->axis[4], usb_gamepad->axis[3], usb_gamepad->hat,
-                                     usb_gamepad->buttons);
-}
-
 #else // !CFG_TUD_HID
 
 static bool grid_usb_hid_ready(void) { return true; }
@@ -93,11 +86,6 @@ static int32_t grid_usb_mouse_move(struct grid_mouse_model* usb_mouse, int8_t po
   (void)usb_mouse;
   (void)position;
   (void)axis;
-  return 0;
-}
-
-static int32_t grid_usb_gamepad_report_send(struct grid_gamepad_model* usb_gamepad) {
-  (void)usb_gamepad;
   return 0;
 }
 
@@ -212,15 +200,6 @@ void grid_usb_keyboard_init(struct grid_keyboard_model* usb_keyboard) {
 
 void grid_usb_mouse_init(struct grid_mouse_model* usb_mouse) { usb_mouse->buttons = 0; }
 
-void grid_usb_gamepad_init(struct grid_gamepad_model* usb_gamepad, uint16_t buffer_size) {
-  usb_gamepad->buttons = 0;
-  usb_gamepad->hat = 0;
-  for (uint8_t i = 0; i < GAMEPAD_AXIS_COUNT; i++) {
-    usb_gamepad->axis[i] = 0;
-  }
-  assert(grid_swsr_malloc(&usb_gamepad->tx, buffer_size) == 0);
-}
-
 void grid_usb_macro_init(struct grid_macro_model* usb_macro, uint16_t buffer_size, struct grid_keyboard_model* usb_keyboard, struct grid_mouse_model* usb_mouse) {
 
   usb_macro->tx_rtc_lasttimestamp = grid_platform_rtc_get_micros();
@@ -235,17 +214,15 @@ void grid_usb_macro_init(struct grid_macro_model* usb_macro, uint16_t buffer_siz
 // Connect / disconnect
 //--------------------------------------------------------------------+
 
-void grid_usb_hid_on_connect(struct grid_macro_model* usb_macro, struct grid_gamepad_model* usb_gamepad, struct grid_keyboard_model* usb_keyboard, struct grid_mouse_model* usb_mouse) {
+void grid_usb_hid_on_connect(struct grid_macro_model* usb_macro, struct grid_keyboard_model* usb_keyboard, struct grid_mouse_model* usb_mouse) {
   (void)usb_macro;
-  (void)usb_gamepad;
   (void)usb_keyboard;
   (void)usb_mouse;
 }
 
-void grid_usb_hid_on_disconnect(struct grid_macro_model* usb_macro, struct grid_gamepad_model* usb_gamepad, struct grid_keyboard_model* usb_keyboard, struct grid_mouse_model* usb_mouse) {
+void grid_usb_hid_on_disconnect(struct grid_macro_model* usb_macro, struct grid_keyboard_model* usb_keyboard, struct grid_mouse_model* usb_mouse) {
   grid_swsr_read(&usb_macro->tx, NULL, grid_swsr_size(&usb_macro->tx));
   usb_macro->has_next = false;
-  grid_swsr_read(&usb_gamepad->tx, NULL, grid_swsr_size(&usb_gamepad->tx));
   usb_keyboard->active_key_count = 0;
   usb_mouse->buttons = 0;
 }
@@ -253,71 +230,6 @@ void grid_usb_hid_on_disconnect(struct grid_macro_model* usb_macro, struct grid_
 //--------------------------------------------------------------------+
 // TX push / flush / available
 //--------------------------------------------------------------------+
-
-int32_t grid_usb_gamepad_axis_move(struct grid_gamepad_model* usb_gamepad, uint8_t axis, int32_t value) {
-  (void)usb_gamepad;
-  if (axis >= GAMEPAD_AXIS_COUNT) {
-    return 0;
-  }
-  struct grid_gamepad_event_desc event = {.type = GRID_GAMEPAD_EVENT_AXIS, .index = axis, .value = (uint8_t)(value + 128)};
-  return grid_usb_gamepad_tx_push(&grid_gamepad_state, event);
-}
-
-int32_t grid_usb_gamepad_button_change(struct grid_gamepad_model* usb_gamepad, uint8_t button, uint8_t value) {
-  (void)usb_gamepad;
-  struct grid_gamepad_event_desc event = {.type = GRID_GAMEPAD_EVENT_BUTTON, .index = button, .value = value};
-  return grid_usb_gamepad_tx_push(&grid_gamepad_state, event);
-}
-
-uint8_t grid_usb_gamepad_tx_push(struct grid_gamepad_model* usb_gamepad, struct grid_gamepad_event_desc event) {
-
-  if (!grid_usb_connected()) {
-    usb_gamepad->tx_dropped++;
-    return 1;
-  }
-
-  uint8_t dropped = 0;
-
-  if (!grid_swsr_writable(&usb_gamepad->tx, sizeof(struct grid_gamepad_event_desc))) {
-    grid_swsr_read(&usb_gamepad->tx, NULL, sizeof(struct grid_gamepad_event_desc));
-    usb_gamepad->tx_dropped++;
-    dropped = 1;
-  }
-
-  grid_swsr_write(&usb_gamepad->tx, &event, sizeof(struct grid_gamepad_event_desc));
-
-  return dropped;
-}
-
-bool grid_usb_gamepad_tx_available(struct grid_gamepad_model* usb_gamepad) { return grid_swsr_readable(&usb_gamepad->tx, sizeof(struct grid_gamepad_event_desc)); }
-
-void grid_usb_gamepad_tx_flush(struct grid_gamepad_model* usb_gamepad) {
-
-  if (!grid_swsr_readable(&usb_gamepad->tx, sizeof(struct grid_gamepad_event_desc))) {
-    return;
-  }
-
-  if (!grid_usb_hid_ready()) {
-    return;
-  }
-
-  struct grid_gamepad_event_desc event;
-  grid_swsr_read(&usb_gamepad->tx, &event, sizeof(struct grid_gamepad_event_desc));
-
-  if (event.type == GRID_GAMEPAD_EVENT_AXIS) {
-    if (event.index < GAMEPAD_AXIS_COUNT) {
-      usb_gamepad->axis[event.index] = (int8_t)(event.value - 128);
-    }
-  } else if (event.type == GRID_GAMEPAD_EVENT_BUTTON) {
-    if (event.value) {
-      usb_gamepad->buttons |= (1u << event.index);
-    } else {
-      usb_gamepad->buttons &= ~(1u << event.index);
-    }
-  }
-
-  assert(grid_usb_gamepad_report_send(usb_gamepad) == 0);
-}
 
 uint8_t grid_usb_macro_tx_push(struct grid_macro_model* usb_macro, struct grid_macro_event_desc event) {
