+++ d51n20a/gcc/Makefile
@@ -92,6 +92,7 @@ SRCS = \
 	common/src/c/grid_usb.c \
 	common/src/c/grid_usb_acm.c \
 	common/src/c/grid_usb_midi.c \
+	common/src/c/grid_usb_gamepad.c \
 	common/src/c/grid_usb_hid.c \
 	common/src/c/grid_utask.c \
 	common/dep/lua-5.5.0/src/dirent.c \
